pyproject.toml
[project]
name = "livingworld"
version = "0.1.0"
description = "Interactive story generator with AI-powered NPCs using Ollama and PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama",
    "langchain",
    "langchain-ollama",
    "langchain-community",
    "asyncpg",
    "sentence-transformers",
    "sqlalchemy",
    "pgvector",
    "python-dotenv",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
]

[project.scripts]
livingworld = "livingworld.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
